# libs/cli-kontrol-api/specs/api.yaml
openapi: 3.0.0
info:
  title: CLI/Kloud-Kontrol API
  version: 1.0.0
paths:
  /flow/create:
    post:
      requestBody:
        description: Create a dev flow
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DevFlowSpec"
      responses:
        "200":
          description: Dev flow creation status
          content:
            application/json:
              schema:
                type: string
  /flow/delete:
    post:
      requestBody:
        description: Delete dev flow (revert back to prod only)
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProdFlowSpec"
      responses:
        "200":
          description: Dev flow creation status
          content:
            application/json:
              schema:
                type: string
  /deploy:
    post:
      requestBody:
        description: Deploy a prod only cluster
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProdFlowSpec"
      responses:
        "200":
          description: Dev flow creation status
          content:
            application/json:
              schema:
                type: string

components:
  schemas:
    ProdFlowSpec:
      type: object
      properties:
        docker-compose:
          type: array
          items:
            x-go-type: compose.ServiceConfig
            x-go-type-import:
              path: github.com/compose-spec/compose-go/types
              name: compose
    DevFlowSpec:
      type: object
      properties:
        image-locator:
          type: string
          example: backend-a:latest
        service-name:
          type: string
          example: backend-service-a
        docker-compose:
          type: array
          items:
            x-go-type: compose.ServiceConfig
            x-go-type-import:
              path: github.com/compose-spec/compose-go/types
              name: compose